[autofill][refactor] Refactor Card Save Methods In CardSaveManager Test
Refactor CreditCardSaveManagerTest to use a mocked version of
ConfirmSaveCreditCardLocally and ConfirmSaveCreditCardToCloud. This
allows for more flexibility and detail in what is being tested (i.e.
tracks the call count, call parameters, callback behaviour) and
reduces the code in TestPaymentsAutofillClient which was only used
for CreditCardSaveManagerTest.

Bug: 40937065,b/346328050
Change-Id: I1c8775d01357e7962053a9c14c6916df9e8c0ebc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5667028
Commit-Queue: Darwin Yang <darwinyang@chromium.org>
Reviewed-by: Olivia Saul <jsaul@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1326269}