Add WakeLock cleanup in destructor
Summary
There are circumstances where a WakeLock is destroyed before it has been deactivated, causing the observers never to be notified of the deactivation.  If an observer is depending on the deactivation behavior, this may cause unexpected issues.  This CL adds cleanup in the WakeLock destructor to call RemoveWakeLock() if it hasn't already been called, resetting the wake_lock_ member and notifying observers.

Background
One instance this happens in product is when a tab is closed while audio is playing.  I believe the reason this is happening is that there is a purposeful delay in registering when audio goes off, as stated in this CL https://chromium-review.googlesource.com/c/chromium/src/+/738198.

When audio is paused,  MediaWebContentsObserver::MaybeUpdateAudibleState() is called twice, and a check is performed each time for recently audible audio.  The first time it is called, audio_stream_monitor->WasRecentlyAudible() returns true, triggering another request for a wake lock (which has no effect).  The second time it is called, audio_stream_monitor->WasRecentlyAudible() returns false, which triggers a request to cancel the wake lock. However, in the case that the audio has been stopped by the tab closing rather than the user press 'pause', it appears that there is only time for the first call to   MediaWebContentsObserver::MaybeUpdateAudibleState().  The result is that the WakeLock is destroyed before it can be removed and observers are not notified of deactivation. (The PowerSaveBlocker created by the WakeLock is destroyed when the WakeLock is destroyed, so in terms of unnecessarily preventing screen dimming, etc there does not appear to be an impact.)

Additionally, adding CHECK(wake_lock_ == nullptr) in the WakeLock destructor without handling that possibility beforehand causes multiple test failures on x64 builds (though not x86), indicating that there may be other race conditions contributing to this issue other than the one mentioned above, at least in testing conditions.

This CL
- Adds cleanup in the WakeLock destructor to notify observers if necessary
- Adds additional cleanup in the WakeLockProvider destructor to cover a race condition triggered in linux-chromeos-rel browsertests where WakeLockProvider was destroyed before WakeLockProvider::OnConnectionError had been called for still-active wake locks
- Adds a feature flag in case the changes need to be switched off

Bug: 349860794
Change-Id: I60147bcadfa8e4b84a3a527b76c8f456b4afadbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5668023
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Sergii Bykov <sbykov@google.com>
Commit-Queue: Sarah Standish <sstandish@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1326283}