Add option to allow IconLabelBubbleView to always paint a background
This CL adds a BackgroundVisibility enum to IconLabelBubbleView that
allows clients force the icon label to always paint a background.

This extends current functionality that only allows for painting
a background while the label is showing.

This CL is purely a refactor and there should be no behavioral
changes. The always-paint-background option is leveraged in the
follow-up CL.

Bug: 352087354
Change-Id: I1e2aaa7c50e0a3e429a99057f07aae3642a485b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5688915
Reviewed-by: David Pennington <dpenning@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326287}