Fix network process crash in IpcDesktopEnvironmentFactory
This crash occurs when the host is restarted while there is an
active connection. The problem is that the DesktopSessionManager
remote is bound to the network task runner but is destroyed on the
main (UI) task runner.

Since the crash is occurring in the AssociatedRemote, a reasonable
fix is to make it a * and delete it on the proper task runner
when IpcDesktopEnvironmentFactory is destroyed.

I also renamed caller_task_runner to network_task_runner because
'caller' implies that it is the task runner which the instance was
created on but it's actually created on the UI task runner and the
'network' task runner is what is actually passed in.

Bug: 350726507
Change-Id: I2d947eb94b900d3f0f24a02abf5d23cb1a4cf63f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5696325
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326333}