Prevent pulsing block animation hanging after screen rotation
In tablet mode, for performing a screen rotation the animations are set
to be ZERO_DURATION.  In the case  for animations that loop, such as the
PulsingBlocks, this can lead to the animation hanging indefinitely and
eventually crashing.  To deal with this edge case, if a delay timer is
fired and there is a duration multiplier for ZERO, restart the timer to
attempt to schedule the animation again.

Bug: b:345364450
Change-Id: Ic03d96e870d6aee07f1969a8c7aaa4f0ded6617e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5690067
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326336}