[SBER Deprecation] Create a new accessor method for CSBRRs
CSBRR, a telemetry report sent from Chrome to Safe Browsing, is
primarily generated by users with SBER enabled. Removing the SBER
dependency would significantly decrease CSBRR volume. To avoid this, a
new accessor method has been added to return the SBER pref value, which
the ping manager uses to decide whether to send CSBRR. This is the part
of the phase one implementation of the SBER deprecation, and this method
will be removed once the mitigation is implemented.


Bug: 336547987
Change-Id: Icb803dd68dd9c4f96408b53e3aa683d33e973d5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5661058
Commit-Queue: Zack Han <zackhan@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326340}