When moving elements around in circular_deque, use modulo arithmetic
When we copy elements to make room for new ones, it's possible that
we will copy from a range, or to a range, that cross the back of the
underlying buffer and wrap around to the start. As we move things in
reverse, when we subtract we're wrapping around to an invalid address
instead of to the back of the buffer. Fix this by applying the capacity
as the modulus.

This OOB should have been caught by bounds checks in VectorBuffer but
they were defeated by our unfortunately-very-likely foe: integer
overflow. Move the UNSAFE_BUFFERS() to as_span() which does not do
any arithmetic with arguments and implement subspan() through base::span
which correctly handles overflow.

R=dcheng@chromium.org

Fuchsia-Binary-Size: PGO size artifacts
Fixed: 352481077
Change-Id: I52e0e882b71eb54d791752a9dab6ddc110d60295
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5695465
Auto-Submit: danakj <danakj@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326431}