Fixing a crash when the Daemon process is stopped
This crash is occurring because the CrashFileUploader spawns a new
thread which it uses for uploading the minidumps, the problem is
that this binds the remote in TransitionalURLLoaderFactoryOwner to
that thread and when it comes time to tear down the instance, the
remote CHECKs and crashes because it wasn't destroyed on the thread
pool thread.

Though I reproduced the crash on Windows, I suspect this was also a
problem on Linux, though the uploader is not configured to report
crashes.

I fixed this by introducing a new class which encapsulates the
various loaders and such needed to watch for and upload minidumps.
This class will execute these operations on the thread it was created
on. For the Windows use case, I create a SequenceBound instance and
for Linux, I just use the existing thread that executes main().

Bug: 350722689
Change-Id: I3c6319481ea4c2b87336615626b8364a58215de8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5698208
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326468}