Add in dialog to save the conversation for Sparky
A new field of dialog will be used to store the conversation and history
of actions. It will also allow for multiple actions within one dialog
response. This CL adds in the new message and struct and include some
util functions to help process the struct.

Bug: 351099209
Test: Tested in unit tests within SparkyUtilTest
Change-Id: I59588a79443c68bf2ec03232348a12114b79af86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5689768
Commit-Queue: Lauren Commeignes <laurencom@chromium.org>
Reviewed-by: Xinglong Luan <alanlxl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326470}