[CodeHealth] Remove RAW_PTR_EXCLUSION for 'ExtensionViewViews::container_'
This replaces the container_ pointer in ExtensionViewViews with a
raw_ptr. The raw_ptr is preferable over a naked pointer because it is
able to detect dangling pointers and prevent UAF.

raw_ptr should be safe in this case, because OnPropertyChanged in
ExtensionViewViews uses the field address of the container_ property.
This read should not be affected by the use of raw_ptr.

No perf test for this change because it is strictly browser only code.

Fixed: 346693592
Change-Id: I9467e427cdc72ececcc9cc5bf6c0c5aada0e8521
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5691120
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326491}