TabDragging tests: Override fallback tab dragging field trial
https://crrev.com/c/5665167 added a field trial to enable the
kAllowWindowDragUsingSystemDragDrop feature. This means that the feature
is always enabled in tests. For the TabDragging tests, it should only be
enabled if the test parameter indicates it.

This accidental enablement caused the one test that still was failing
with fallback tab dragging, DeleteTabsWhileDetached, to be disabled in
https://crrev.com/c/5671897. With the changes from this CL, and also
thanks to https://crrev.com/c/5538283 fixing the failure with fallback
tab dragging, we can enable it again.

Bug: 40238145
Change-Id: I5101c5da16e81a4a9cde1fcd5ea71e65355af45d
Fixed: 350621332
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5691787
Commit-Queue: Max Ihlenfeldt <max@igalia.com>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326558}