[views-ax] Migrating kExpanded and kCollapsed attribute to new system for combobox view
This CL migrates the kExpanded and kCollapsed state in
ui\views\controls\combobox\combobox.cc whenever its value should
change, rather than querying the value and computing it only when needed.

This CL also contains unit tests for the view testing
the accessibility states when the view is open and closed.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: Ibc51ad9a8d0ea3f7dbe705f0c8924103414b6b37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5653576
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1326584}