[WebAuthn] Make ChangePinControllerImpl per document
This fixes a bug when user navigates away and re-visits the PWM settings
and clicks Change PIN. Nothing was happening because the rfh for the
previous PWM settings is gone and AuthenticatorRequestDialogModel
depends on it.

Also ChangePinController is not required if the user navigates away from
the PWM settings UI. It can be recreated.

Bug: b/352309219
Change-Id: I83fa2de5ca3bac9d18ae99ec9d2199d1ae2d63ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5694707
Commit-Queue: Adem Derinel <derinel@google.com>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326587}