[ios] Rename BookmarkModelType to BookmarkStorageType
This patch tackles a TODO to avoid the outdated terminology and
mindset of dealing with two bookmark models. A close equivalent, and
a terminology with plenty of precedents in the codebase, is to use
"storage type" to distinguish whether the code is dealing with local
vs account bookmarks.

In the long run, it is conceivable that components/bookmarks will
standardize this notion, but meanwhile the iOS-specific enum is used
with a slightly better name.

Bug: 346918509
Change-Id: Ie9e35fd7b14bc642c61863876263ca971fe23a88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5701225
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326646}