[views-ax] Migrating kExpanded and kCollapsed attribute to new system for suggestions section
This CL migrates the kExpanded and kCollapsed state in
ash\system\holding_space\suggestions_section.cc whenever its value 
should change, rather than querying the value and computing it only when needed.

This CL also contains unit tests for the view testing
the accessibility states when the view is expanded and collapsed.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: Ide8065aa9e926b6926a2b53338981a4e403ce452
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5654831
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1326653}