[Magic Stack] Making Segmentation API call ondemand.
Enabling ondemand option for Segmentation API with the ability to cache result and use the cached result if ondemand result is unavailable for Clank Magic Stack. Ondemand is required to make sure that the model is able to use freshness inputs passed as input context.

Bug: b/337878177
Change-Id: I5a43d5e9ec7b259a71f707890c3a7ee21d41a3da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5665724
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Ritika Gupta <ritikagup@google.com>
Reviewed-by: Siddhartha S <ssid@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326669}