[What's New] Fix issues with TimeToLoadContent metric
This metric is being recorded multiple times for each page load, due to
the redirect in the iframe. Additionally, when reloading the page, the
previous navigation_start_time_ was being used to calculate an incorrect
value for the load time.

Bug: 350455493, 350455131
Change-Id: Ibc810312e95be6f57b49b26e6e9d387d33bc61a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5670302
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Mickey Burks <mickeyburks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326685}