[UMA] Add auth error bucket to UMA.ProfileSignInStatusV2 and UMA.ProfileSyncStatusV2
The new bucket is populated for UMA.ProfileSignInStatusV2, but not
for UMA.ProfileSyncStatusV2 (to avoid changing the meaning of the
histogram).

It is fine to add the bucket to UMA.ProfileSignInStatusV2 without
renaming the histogram, because the "signin pending" state is new.
Existing clients would never have recorded it.

Bug: 344764060
Change-Id: I32ac9b5e774222e9343a4e639301a77b3248ac78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5630302
Reviewed-by: Victor Vianna <victorvianna@google.com>
Commit-Queue: David Roger <droger@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1326695}