[iOS] Filter identities according to the profile
When multiple profiles are enabled, the default profile sees only
non-managed identities, and other secondary profiles see only managed
identities.

The test to know if an identity is managed or not is based on
"@gmail.com" or not.
Ideally, this should be done using GCRSSOService, but the async APIs
cannot be used in this sync implementation.
And cache APIs might trigger flakiness if the cache disappear on the
next cold start.

Low-Coverage-Reason: EXPERIMENTAL_CODE This code is used only when multiprofile is enabled.
Bug: 331783685
Change-Id: I97f78994b63c0c6e4a34f130987b73d760245c47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5688831
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326702}