[SHoA] Add an API to determine the module states
Refactor out the logic to determine the state of each module into an API
that can be re-used to determine the overall browser state ("all good"
state if all ModuleOptions have a ModuleState >= INFO) and the order
of the modules in the SH UI (WARNING states should be ordered first in
the preference list).

This adds the base logic for determining the module state, the "all good" indicator and ordering will be done in a followup.

Bug: 324562205
Change-Id: I7575fc93d8c9ab0d8d0d06056c4c6a2ca64847e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5692062
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Zaina Al-Mashni <zalmashni@google.com>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326724}