dnt: Refactor NavigationHandler#triggerUi to take edge
This API currently specifies whether the navigation is back or forward
and callers make this decision based on the edge. However, we'd like to
mirror the back edge on devices using an RTL language which would
require each caller to make this determination and adjust the argument
accordingly.

This CL refactors this API to pass in the physical edge being used so
that the shared implementation can determine, based on UI language,
whether the given edge triggers a back or forward navigation.

Bug: 331778964
Change-Id: I3a1adca0ddb3c8517f1662b1157579fd84432a0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5689874
Commit-Queue: David Bokan <bokan@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326734}