[Tab Switcher] Remove TabSwitcherPaneAdapter
With the submission of the following CL this adapter is no longer needed
https://chromium-review.googlesource.com/c/chromium/src/+/5696286

This class was a proxy for TabSwitcher calls to TabSwitcherPaneBase. It
was used to reduce the API surface of the TabSwitcher for Hub. Now that
there is a 1:1 relationship between the TabSwitcher API and the
TabSwitcherPaneBase it is no longer necessary.

Bug: 352536546
Change-Id: I19d33a3e3074e11849ad7df4361976a6f57b6a5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5702596
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326746}