IDB: Don't create ClientStateChecker when BucketContext is invalid
The linked crash is happening because of a call to
`DedicatedWorkerDevToolsAgentHost::GetFor()` (from
`DedicatedWorkerHost::GetDevToolsToken()`) when `PlzDedicatedWorker` is
disabled, which is an unsupported scenario.

`DedicatedWorkerHost::GetDevToolsToken()` is actually never expected to
be called because `DedicatedWorkerHost::GetAssociatedRenderFrameHostId()`
always returns the ancestor RFH, which means a
`DocumentIndexedDBClientStateChecker` instance is supposed to be used
for it (which returns the DevTools token from the RFH instead).

However, a `DedicatedWorkerHost` can momentarily outlive its owning RFH
- https://issues.chromium.org/issues/40807127. When a call to
`IndexedDBClientStateCheckerFactory::InitializePendingRemote()` happens
at such a time, the `RenderFrameHost::FromID()` call returns nullptr
and we end up using a `NoDocumentIndexedDBClientStateChecker` instance.

When this happens, we call `DedicatedWorkerHost::GetDevToolsToken()` to
pass to `NoDocumentIndexedDBClientStateChecker`, which fails when
`PlzDedicatedWorker` is disabled.

Hence, this CL:
- Updates `InitializePendingRemote()` to not bind the remote instead
when the above scenario occurs, and
- Updates `DedicatedWorkerHost::GetDevToolsToken()` to `CHECK` that
`PlzDedicatedWorker` is enabled.

Bug: 349831508
Change-Id: I77f84357908e3c170965851c8cd0e579eb300c9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5670091
Auto-Submit: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Commit-Queue: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326762}